Python and HTML update
I made some additions to the Python code. It now functions very
similarly to our JavaScript code (without the HTML hookin yet). The
script now successfully collects all user data, project data, and update
data; and organizes that data throughout a series of logical objects.
Note: the code takes a while to run!

Also, I added a very simple and quick message to the bottom of the
webpage asking for user feedback. This message can (and should) be
adjusted with ease.