Temporary fix
this file contains the quick changes I've made to make the Giddy TV operational for the projector. Note that the code for filtering users (users within past 7 days) has been updated but commented out (I'm currently the only person to have made a profile within the past week apparently). This means running the app will try to gather data on all 1000+ users -- expect it to take some time. I also removed instances in which we add "/raw" to image URLs. Images no longer need that check. Finally, I removed code which checked for "mp4" files. In the future this can be troublesome, but at the moment, it works (absolutely NO update is currently in Giddy).